Added basic classes needed for communication
The server will run a ServerThread serving incoming connections,
assigning Responders to each connection. The client will use a single
Responder to communicate with the server. The Responders allow sending
Packets over the connection and forward incoming packets to
PacketProcessors.
The client and the server will implement their own PacketProcessors.